# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Time Capsule services",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/historical/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new historical event with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HistoricalEvents"
                ],
                "summary": "Add a new historical event",
                "parameters": [
                    {
                        "description": "Add historical event request",
                        "name": "AddHistoricalEvent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.AddHistoricalEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historical event successfully added",
                        "schema": {
                            "$ref": "#/definitions/genproto.AddHistoricalEventResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/historical/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a historical event based on the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HistoricalEvents"
                ],
                "summary": "Delete a historical event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Historical event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historical event successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/genproto.DeleteHistoricalEventResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Historical event not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/historical/gets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all historical events based on provided filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HistoricalEvents"
                ],
                "summary": "Get all historical events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit number of events",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date filter",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historical events found",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetAllHistoricalResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/media/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add media to a specific memory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Add new media",
                "parameters": [
                    {
                        "description": "Add media request",
                        "name": "AddMedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.AddMediaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media successfully added",
                        "schema": {
                            "$ref": "#/definitions/genproto.AddMediaResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/media/delete/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete media by memory ID and media ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Delete media",
                "parameters": [
                    {
                        "description": "Delete media request",
                        "name": "DeleteMedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.DeleteMediaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/genproto.DeleteMediaResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/media/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all media related to a specific memory ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get media by memory ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get media by memory ID request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetMediaByMemoryIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/media/gets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all media, optionally filtered by type and URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get all media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "URL",
                        "name": "url",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetAllMediaResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/media/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update media details such as type and URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Update media",
                "parameters": [
                    {
                        "description": "Update media request",
                        "name": "UpdateMedia",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateMediaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media updated successfully",
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdateMediaResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genproto.AddHistoricalEventRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "genproto.AddHistoricalEventResponse": {
            "type": "object"
        },
        "genproto.AddMediaRequest": {
            "type": "object",
            "properties": {
                "media_id": {
                    "type": "string"
                },
                "memory_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "genproto.AddMediaResponse": {
            "type": "object"
        },
...
        "genproto.DeleteHistoricalEventResponse": {
            "type": "object"
        },
        "genproto.DeleteMediaRequest": {
            "type": "object",
            "properties": {
                "media_id": {
                    "type": "string"
                },
                "memory_id": {
                    "type": "string"
                }
            }
        },
        "genproto.DeleteMediaResponse": {
            "type": "object"
        },
...
        "genproto.GetAllHistoricalResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.HistoricalEvents"
                    }
                }
            }
        },
        "genproto.GetAllMediaResponse": {
            "type": "object",
            "properties": {
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.Media"
                    }
                }
            }
        },
...
        "genproto.GetMediaByMemoryIdResponse": {
            "type": "object",
            "properties": {
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.Media"
                    }
                }
            }
        },
...
        "genproto.HistoricalEvents": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "genproto.Media": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "memory_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "genproto.UpdateMediaRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "memory_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "genproto.UpdateMediaResponse": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}